+++ src/main/groovy/fr/nereide/reference/xml/provider/ServiceEngineReferenceProvider.groovy
@@ -1,0 +1,18 @@
+package fr.nereide.reference.xml.provider
+
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiReference
+import com.intellij.psi.PsiReferenceProvider
+import com.intellij.util.ProcessingContext
+import fr.nereide.project.PluginActivator
+import fr.nereide.reference.xml.EngineReference
+import org.jetbrains.annotations.NotNull
+
+class ServiceEngineReferenceProvider extends PsiReferenceProvider {
+
+    PsiReference[] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
+        if (!PluginActivator.getInstance(element.project).isActive()) return []
+        EngineReference ref = new EngineReference(element)
+        return ref ? ref as PsiReference[] : PsiReference.EMPTY_ARRAY
+    }
+}